pyproject.toml
[tool.poetry]
name = "slides-project"
version = "0.1.0"
description = "generate slides with langchain"
authors = ["speters9 <speters9@stanford.edu>"]
readme = "README.md"
#packages = [{include = "slides_project"}]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pandas = "^2.2.2"
scikit-learn = "^1.4.2"
arrow = "^1.3.0"
numpy = "^1.26.4"
python-dotenv = "^1.0.1"
transformers = "^4.40.1"
langchainhub = "^0.1.15"
langchain = "^0.2.9"
langchain-community = "^0.2.7"
sentence-transformers = "^3.0.1"
langchain-huggingface = "^0.0.3"
openai = "^1.37.0"
langchain-openai = "^0.1.17"
pypdf2 = "^3.0.1"
torch = { version = "2.0.0", source="torch"}
torchvision = { version = "0.15.1", source="torch"}
torchaudio = { version = "2.0.1", source="torch"}
python-docx = "^1.1.2"
pytesseract = "^0.3.10"
pdf2image = "^1.17.0"
pillow = "^10.4.0"
pymupdf = "^1.24.9"
tiktoken = "^0.7.0"
pyprojroot = "^0.3.0"


[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu118"
priority = "supplemental"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
